Move Maven test utilities to their own package
Given the additional test specifications in the `maven` package, this should better logically distinguish test utility classes from test specifications.
