+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/maven/UnusualProjectStructureTest.groovy
@@ -1,6 +1,8 @@
 package nu.studer.teamcity.buildscan.agent.maven
 
 import nu.studer.teamcity.buildscan.agent.TcPluginConfig
+import nu.studer.teamcity.buildscan.agent.maven.testutils.MavenBuildStepConfig
+import nu.studer.teamcity.buildscan.agent.maven.testutils.MavenProject
 
 import static org.junit.Assume.assumeTrue
