+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/maven/testutils/GroupArtifactVersion.groovy
@@ -1,4 +1,4 @@
-package nu.studer.teamcity.buildscan.agent.maven
+package nu.studer.teamcity.buildscan.agent.maven.testutils
 
 final class GroupArtifactVersion {
